Move UnrealResourcePreparer to a separate file and rename it.
